+++ R/datasets.R
@@ -2599,6 +2599,8 @@ NULL
 #'        nrow=4, ncol=3)
 #' # assign these to the formula factor
 #' contrasts(dogfood$formula) <- C
+#' # re-fit the model
+#' dogfood.mod <- lm(cbind(start, amount) ~ formula, data=dogfood)
 #' 
 #' dogfood.mod <- lm(cbind(start, amount) ~ formula, data=dogfood)
 #' Anova(dogfood.mod)